Move FiltersDropdown into components folder
