+++ src/pages/LoansPage/components/MyLoansList/MyLoansList.tsx
@@ -6,7 +6,7 @@ import { useOnClickOutside } from '@frakt/hooks';
 import { ConnectWalletSection } from '../../../../components/ConnectWalletSection';
 import FiltersDropdown, {
   useFiltersModal,
-} from '../../../../componentsNew/FiltersDropdown';
+} from '../../../../components/FiltersDropdown';
 import { LoansList } from '../LoansList';
 import { selectTotalDebt } from '../../../../state/loans/selectors';
 import Button from '../../../../components/Button';
